requirements.txt
requests == 2.21.0
mysql-connector == 2.1.6
